fix(quantic): improved condition that indicates when to display timeframe facet  (#5143)
## [SFINT-6077](https://coveord.atlassian.net/browse/SFINT-6077)

## 🛠️ Summary

This PR addresses the issue described in [Jira: Quantic Timeframe Facet
filter input is always
displayed](https://coveord.atlassian.net/browse/SFINT-6077), where the
**timeframe facet** displays the date picker inputs even when it doesn’t
contain any facet values. This leads to unnecessary UI elements being
shown when the facet field is irrelevant due to applied filters.



https://github.com/user-attachments/assets/735d4deb-18ac-4a79-bcdb-dba0fcefb0f1




## ✅ Expected Behavior

When the `with-date-picker` property is set to `true`, the date filter
inputs **should only be displayed when the filter is actually applicable
and can filter the current result set**.



## ✨ Implementation

Introduced a new computed property to determine when to show the date
picker:

```ts
/**
 * Indicates whether to display the date picker.
 */
get shouldDisplayDatePicker() {
  if (!this.withDatePicker) {
    return false;
  }
  if (this.dateFilterState?.range) {
    return true;
  }
  if (!this.hasResults) {
    return false;
  }
  const atLeastOneValueWithResultsOrActive = this.facetState?.values.some(
    (value) => value.numberOfResults || value.state === 'selected'
  );
  return atLeastOneValueWithResultsOrActive;
}
```

This ensures the date picker is only shown when:

- `withDatePicker` is enabled  
**AND** at least one of the following is true:
  - A date range filter is already selected (`dateFilterState.range`)
  - The facet has results, and at least one value:
    - Has results (`numberOfResults > 0`)
    - **OR** is selected (`state === 'selected'`)


## 🫡 Demo


https://github.com/user-attachments/assets/a2b20c0a-07dd-4580-92ae-124bdd1ec93e







[SFINT-6077]:
https://coveord.atlassian.net/browse/SFINT-6077?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ