chore(atomic): add a boolean converter for Lit (#5121)
https://coveord.atlassian.net/browse/KIT-3972

By default, with Lit, boolean attributes need to be omitted to be
converted to `false`. `attribute="false"` is considered `true`. This is
different from Stencil where attribute="false" is considered `false`.
This PR adds a boolean converter so that `attribute="false"` gets
converted to `false` and so we don't make breaking changes.

---------

Co-authored-by: ylakhdar <ylakhdar@coveo.com>