fix(atomic-icon): move css to components layer (#5139)
Currently, the stylesheet of the component was unlayered, meaning that
it had a higher priority than all layered stylesheets, which caused the
`hidden` class of the layer class to be overridden.


![image](https://github.com/user-attachments/assets/76ef7d5c-d391-4a56-bc2a-14d55a9aabc2)
https://developer.mozilla.org/en-US/docs/Web/CSS/@layer

To fix that, I put the whole shebang into the components layer, so that
the utilities has priority over it.

IMO, we should have a "sandwich" of CSS layers from the least important
to the most:
1. base: everything that components and elements should default on
2. components: everything specific to components
3. utilities: everything components _can_ use

I don't know if we want to put an e2e for that tho...
UT doesn't have the whole styles loaded from what I experimented with