chore(pipeline): Migrate deployment configuration of ui-kit [DT-7842] (#5111)
# PLEASE MERGE IT YOURSELF
  
Hi there 🤠!

In order to make the config easier to use in the context of the
Multi-Account Framework, we are running 🏃‍️🎽👟 a script to migrate
deployment configurations!

While the change is not breaking 💥 since the revise is done
automatically when you create a package 📦, updating your code will just
make it so that you don't have drift 🏎️ between the config you submit
and what the pipeline actually see 👀.

See [the announcement 📢 in
Confluence](https://coveord.atlassian.net/wiki/spaces/CM/pages/5024317641/Deployment+Config+version+19+Changes+for+Overrides+and+more)
for the details

The script also updates `deployment_config_version` so that it does not
need to run every time 🕛️ you create a package in the future. Please
forgive any formatting changes as those are hard to avoid 🙏.

I hope you accept my little gift 🎁.
  
# PLEASE MERGE IT YOURSELF

Note (dblanchette): Opening this manually because the automated script
breaks on the `$[IS_NIGHTLY]` placeholder

DT-7842