fix(headless): advanced search queries parameters are not passed along (#5173)
[SVCC-4986](https://coveord.atlassian.net/browse/SVCC-4986)

## Bug Description

On Wednesday, April 9th, we received a message about a customer having
issues with the search page. The generated answer was returning
citations that should be filtered out. They open [a
discuss](https://discuss.coveo.com/t/lpl-crga-issues-with-or-without-answer-manager-configuration/12854/5)
with details.

A similar search page as the setup by the customer but using facets
filtering was provided to us to reproduce the issue but we were unable
to reproduce it.

After meeting with customers and observing the behaviour, we discovered
that the customer was leveraging the cq advance search queries parameter
to be able to filter results displayed to user, based on a value set in
the user session.

## Root cause search

Based on the fact that searches were not returning valid results only
when an answerConfigurationId was set in the search page, we started by
making sure that the answerAPI was handling correctly this advanced
search query parameter. Once that was validated, we moved up the chain
and made sure in headless that this parameter was transmitted to
AnswerAPI in the generated request. There, we found out that not only
`cq` parameter was not passed along to the request, but also that the
`aq`, `dq` and `lq` advanced search query parameters were not passed
along.

## Solution

Even though the customer is only using `cq`, we decided to pass along
all advanced search queries parameters to the API.

## Tests & Example page updated

Unit test mock and expected payload have been updated. I also updated
the example genqa page so that we have an example of how-to activate
those parameters nearby.

## How to test locally

* [Make sure your local setup is ready to
build](https://github.com/coveo/ui-kit?tab=readme-ov-file#install) and
[run
atomic](https://github.com/coveo/ui-kit?tab=readme-ov-file#development-mode)
* Pull this branch
* Run headless unit tests with the following command `npx nx run
headless:test`
* Edit `packages/atomic/dev/examples/genqa.html` file to add a
answerConfirgurationId(`fc581be0-6e61-4039-ab26-a3f2f52f308f`) as a
property in the component at line **124**
* Run atomic dev with this command `npx nx run atomic:dev`
* Open your local browser to [this
page](http://localhost:3333/examples/genqa.html#q=how%20to%20resolve%20netflix%20connection%20with%20tivo)
* Open your developer tool and then Network tab
* Look for a `/generate` request
* You should find that the parameters in the request like in the
following screenshot
<img width="888" alt="Screenshot 2025-04-11 at 2 32 46 PM"
src="https://github.com/user-attachments/assets/8a8b900c-c671-4e4b-a28e-2d1e92d19566"
/>


[SVCC-4986]:
https://coveord.atlassian.net/browse/SVCC-4986?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ