chore(atomic): add internal element for rendering search box suggestion elements (#5178)
https://coveord.atlassian.net/browse/KIT-4127


* Replaced `ButtonSearchSuggestion` and `SimpleSearchSuggestion`
components with a new `atomic-suggestion-renderer` component in search
box components.


The reason for this is to isolate the rendering logic of the 'content'
property of the `SearchBoxSuggestionElement` interface. This property
can be Stencil VNode or native Element so there must be a stencil
component to render it. For lit components using this component, they
will use native Elements.