feat(headless, quantic): added logic to automatically fetch insight interface when building an insight engine (#5187)
## [SFINT-6108](https://coveord.atlassian.net/browse/SFINT-6108)

## The Issue
In the Insight use case, the search hub value is derived from the
Insight interface associated with a specific Insight configuration ID.
This requires implementers to manually dispatch the fetchInterface
action by calling the fetch method on the HeadlessInsightInterface
controller. Without this step, the Insight panel defaults to using
default as the search hub, and this value cannot be overridden via the
Quantic or Atomic UI components.

This manual step is often overlooked, which can lead to an incorrect
search hub value and inaccurate analytics reporting.


https://coveord.atlassian.net/wiki/spaces/SFINT/pages/5079695595/Guiding+Custom+Implementation+of+Quantic+Non-Hosted+Insight+Panels#%E2%9A%99%EF%B8%8F-Internal-change---Issue%3A-Incorrect-Search-Hub-Without-Manual-fetch()-Call-in-Non-Hosted-Panel

## The Solution
To ensure that the Insight panel is initialized with the correct search
hub value, we now automatically fetch the Insight interface during
Insight engine creation. This change reduces the implementation burden
on users and helps prevent misconfigurations that could result in
incorrect analytics data.

<img width="1728" alt="Screenshot 2025-04-18 at 9 00 21 AM"
src="https://github.com/user-attachments/assets/a78b1607-01d3-4c20-acce-52d706648660"
/>


[SFINT-6108]:
https://coveord.atlassian.net/browse/SFINT-6108?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ