docs(headless, headless-react): typedoc styling update and nav cookies as functional (#5195)
[KIT-3964](https://coveord.atlassian.net/browse/KIT-3964)

Based on compliance for cookies, an update was made to TypeDoc to allow
for disabling of local storage usage, and is contained in v0.28.1. This
PR addresses the update to v0.28.1, and incorporates the change to
disable local storage usage by typedoc features when functional cookies
are disabled, respecting compliance requirements.

It also updates the styling of typedoc generated sites to better follow
the style present on the main docs site. The header bar now contains the
same functionality as our main docs site and the style of the body is
more consistent with our main docs also.

[KIT-3964]:
https://coveord.atlassian.net/browse/KIT-3964?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Jean-Philippe Marceau <39384459+jpmarceau@users.noreply.github.com>
Co-authored-by: Louis Bompart <lbompart@coveo.com>