chore: add checkbox render function (#5215)
https://coveord.atlassian.net/browse/KIT-4152

Add core deps for facet (checkbox render functions)

## Bonus
Added an identity function `tw` that will throw an error whenever the
tailwind classes are not ordered.
When

### Explaination
By default, when passing classes to an element in a html template, an
prettier error is thrown whenever the class order is not respected

![image](https://github.com/user-attachments/assets/0c8cd2e4-8b60-4b9f-9e89-bc830965dcb5)

However, this check is not applied when the classes are inside an
object.
With this new identify function, we can have the same error 

![image](https://github.com/user-attachments/assets/3a782541-f3ff-4776-b598-b0cf3614f98a)

---------

Co-authored-by: Louis Bompart <lbompart@coveo.com>
Co-authored-by: developer-experience-bot[bot] <91079284+developer-experience-bot[bot]@users.noreply.github.com>