feat(insight): add analytics on copyToClipboard and sendToFeed (#5235)
SFINT-6026
SFINT-5860

Added analytics on copyToClipboard and sendToFeed and add delayed
tooltip on action.

For the tooltip, I didn't make a custom component since after a
discussion with Louis, we want to revisit tooltips as a whole eventually
to use a native solution like popover.
Lucille already approved of this implementation 👀 


![clipboard](https://github.com/user-attachments/assets/8b8a41dd-6224-4e30-bc01-373466b24a49)

![image](https://github.com/user-attachments/assets/4e63f31c-3d5a-4886-aa6f-1ef7de99d0d9)

![image](https://github.com/user-attachments/assets/c27ca1cf-acd7-435b-93ed-517b2e99b41e)