fix(answerAPI): post feedback (#5244)
# BugFix
## JIRA
https://coveord.atlassian.net/browse/SVCC-4958
## Scope
Submitting a CRGA feedback throws in the console. This has to be fixed
## Fix Context
The way [redux toolkit
query](https://redux-toolkit.js.org/rtk-query/overview) handles
mutations is unknown and not really documented

During the initial implementation, the implementer (me) wrongly assumed
that a HTTP post had to be handled with a `builder.mutation`. Seems like
its not the case.
Using `builder.mutation` throws, while using `builder.query` doesn't
throw. I couldn't find why by looking at the documentation or by
debugging directly in the lib.
## How to test
Load the RGA demo and submit a feedback

[simplescreenrecorder-2025-05-01_11.46.46.webm](https://github.com/user-attachments/assets/b4857925-12b0-4c54-bc53-a47578aee786)


## Code coverage
I do not think we really have a way to test this.