chore(ci): add checks to run atomic e2e or not depending on files changed (#5279)
SFINT-6134

New Schema of the CI:


![image](https://github.com/user-attachments/assets/23b458e2-0118-470b-ab58-9f1c22d2868d)



- add test to run atomic e2e or not
- re-add e2e-quantic to merge-queue checks
- added `packages/bueno` as part of the changes that could trigger
running the Quantic e2e tests (because we use Bueno directly in Quantic)

I created a new script to validate that **all** the files changed match
some patterns.
Used this script to test before running any kind of e2e tests for Atomic
if **ALL** the changes are contained within `/packages/quantic` to skip
all of the Atomic e2e and everything that goes with it,

- e2e-atomic-csp-test
- prepare-playwright-atomic
- playwright-atomic-theming
- e2e-atomic-test
- e2e-atomic-screenshots
- e2e-atomic-react-test
- e2e-atomic-react-nextjs-test
- e2e-iife-test : Which doesn't test anything btw???
- e2e-atomic-angular-test
- e2e-atomic-vuejs-test
- e2e-atomic-stencil-test
- playwright-atomic-hosted-page-test
- playwright-headless-ssr-commerce
- e2e-atomic-insight-panel-test

And I also re-added the `e2e-quantic` as part of the
`is-valid-merge-queue` checks.

See the comments for the manual tests performed, but TL;DR

quantic-only changes => https://github.com/coveo/ui-kit/pull/5287
atomic-on changes => https://github.com/coveo/ui-kit/pull/5299
both atomic and quantic (because I changed package.json at root) =>
https://github.com/coveo/ui-kit/pull/5298