fix(atomic): fix atomic-recs-result-template rendering error (#5313)
https://coveord.atlassian.net/browse/KIT-4224

The `atomic-recs-result-template` was throwing errors in Angular because
its host element's parent wasn't defined when we were creating the
`ResultTemplateCommon` instance in the constructor.

I applied the same fix to `atomic-recs-result-template` that was applied
to `atomic-result-template` here:
https://github.com/coveo/ui-kit/pull/5070/files#diff-869c29501cc46796f3035d2f9e0b3d564509824dd998c0c24d49efe4f2dd0423,
i.e., creating he `ResultTemplateCommon` instance in the component's
`connectedCallback` instead of in the constructor.

I also included a recommendation interface in the Angular sample.

For good measure, I preemptively applied the fix to other result
template components that create `ResultTemplateCommon` instances
(`atomic-insight-result-children-template`,
`atomic-result-children-template`).