chore(headless,bueno): externalize all packages on node outputs (#4921)
A couple of changes to allow source maps to work properly for our users,
and avoid package duplications in bundles:
- Externalize all dependencies for NPM output
Stop bundling on NPM. This was an error, causing duplication of heavy
libraries such as Redux on customer websites, reducing their overall
performance.
- Merge back conditional exports: we don't need those anymore since we
do not pre-bundle anymore and the code in our package is isomorphic.
- Create a new CDN output that keeps on bundling: Used for our
static.cloud.coveo.com shenanigans
    
- Stencil sample required stabilization due to its known deficiencies
with modern module resolution.
- Sadly, we cant change module resolve conditions programmatically yet
in a given node process, thus the horrible `spawnSync` hack.
- Even the hack cannot handle/overwrite/discard `node`, so special
resolution for this one --> require NodeJS 22.14.0 for
[`module.findPackageJSON`](https://nodejs.org/api/module.html#modulefindpackagejsonspecifier-base)

------

https://coveord.atlassian.net/browse/KIT-4148

---------

Co-authored-by: GitHub Actions Bot <>