fix(atomic): refactor smart snippets to ensure their internally used custom elements are defined in atomic angular (#5367)
https://coveord.atlassian.net/browse/KIT-4233

## Issue

At the moment, when running a project relying on Atomic Angular, the
`atomic-smart-snippet` and `atomic-smart-snippet-suggestions` components
do not work; they render empty containers:

![Screenshot 2025-05-30 at 09 02
19](https://github.com/user-attachments/assets/9a2c9232-bd43-4460-8b45-d852d66d04e1)

## Causes

There are two causes:

1. We are no longer including our lazy loader in the Atomic Angular
module. It was removed
[here](https://github.com/coveo/ui-kit/pull/5070/files#diff-0798d397c8638f343629aab4b0a7f18e1ba1cfdbf4d77e37883dfdcc3198cdb3)
because it was not working for Lit components. It has been replaced by
logic in build-lit.mjs. By all accounts this change is legitimate and we
do not want to revert it. However, it means that "undeclared" components
(such as the ones used internally by `atomic-smart-snippet` and
`atomic-smart-snippet-suggestions`, e.g.,
`atomic-smart-snippet-expandable-answer` or
`atomic-smart-snippet-source` will not be defined.

2. The render functions of the affected components are calling `render`
methods on "common" classes. This is overengineered and prevents Stencil
from detecting the custom elements that the affected components depend
on.

## Solution

This PR simplifies the render `atomic-smart-snippet`,
`atomic-insight-smart-snippet`, `atomic-smart-snippet-suggestions`, and
`atomic-insight-smart-snippet-suggestions` render functions by
refactoring the "common" classes into sets of functional components, and
bringing back the rendering logic inside each component.

If a custom element is directly and unconditionally referenced within a
component's render function, Stencil will necessarily detect it as a
dependency of that component and ensure that it is defined.

This can be verified by looking at the Stencil-generated
`dist/atomic/components/atomic-smart-snippet.js` file, for instance.

Before this PR, none of the component's dependencies are defined:

![Screenshot 2025-05-30 at 09 33
49](https://github.com/user-attachments/assets/2f835cd1-65c4-4cbb-8978-4e910e0bab50)

After this PR, they are defined:

![Screenshot 2025-05-30 at 09 37
39](https://github.com/user-attachments/assets/f906a2a7-dfde-4615-884f-4c7bb0cb297c)

And we can see that the components now work correctly in Atomic Angular:

![Screenshot 2025-05-30 at 09 47
59](https://github.com/user-attachments/assets/d8415a62-fb4a-4b7a-9798-8da7e19753b7)

## Testing

To test the `atomic-smart-snippet` and
`atomic-smart-snippet-suggestions` fixes in Atomic Angular:

1. `npm run build`
2. `npm run dev -w=@coveo/atomic-angular-samples`
3.  `open http://localhost:4200/`
4. Click "Samples search interface with Atomic Angular
(@coveo/atomic-angular)"
5. In the search box, enter `query data lake` and submit the query
6. Bask in the glory

To ensure that `atomic-insight-smart-snippet` and
`atomic-insight-smart-snippet-suggestions` are unaffected by the change:

1. `npm run build`
2. `npx nx run atomic:dev`
3. `open http://localhost:3333`
4. Click the "Insight Panel" tab (or stay on the main tab if you want to
test that the non-insight versions of the components are also unaffected
by the change)
5. In the search box, enter `query data lake` and submit the query
6. Bask in the glory