ci: run only github actions affected by PR (#5402)
https://coveord.atlassian.net/browse/KIT-4301

Using a NX command, we can determine what projects are affected by the
files changed. We can then add a condition to every github action that
will check if one of its dependant project is affected.

Example :

This typedoc action builds typedoc for headless & headless-react. If one
of those is in the projects affected, it will run it. If not, it will
skip it.

```
  typedoc:
    name: 'Build typedoc'
    needs: affected
    if: contains(needs.affected.outputs.projects, 'headless') || contains(needs.affected.outputs.projects, 'headless-react')
```


**Please review each action to make sure it has the right projects in
the condition.**