fix(headless-ssr-react): ensure provider's children are not remounted on hydration (#5404)
The current provider has a fundamental flaw: When we change the parent
element where the children are rendered, the children are remounted.

To solve this, I mashed up the static & hydrated provider together
essentially.
And while at it, I reckon we should deprecate the hydrated & static
because they alone are a bit useless IMO compared to the "isomorphic"
one.

---------

Co-authored-by: Alex Prudhomme <78121423+alexprudhomme@users.noreply.github.com>