fix(headless): respect query param tab on page refresh (#5412)
https://coveord.atlassian.net/browse/KIT-4289

The initial refresh of the tab does not filter out the right and wrong
facet to send to the API. This makes it so queries responses are
different when you first 'arrive' on a page with that tab selected VS
when you click on that tab from another one.

I think this was due to this redux best practice :
https://stackoverflow.com/questions/67037216/redux-sending-two-dispatch-after-each-other-then-the-first-dispatch-get-lost

It works properly when using the restoreSearchParameters action and
there is no reason why it shouldn't.

I am unsure why this was done before and what was not working
[then](https://github.com/coveo/ui-kit/pull/4944) as there is no
reproducible steps.