chore(headless): stop bundling ESM headless for NPM (#5346)
Bundling packages on NPM is "not good" as it prevents consumers from
properly treeshaking them.

This PR brings treeshaking for Headless for its ESM exports.

## Why not CJS?
With `require(esm)` around the corner, I have good hope of putting the 🪓
in our CJS output in v4, with minimal customer impact (i.e. they can
either update to Node 24 that'll be LTS and will _likely_ include Stable
`require(esm)`, or switch to ESM).

### What if we need CJS
Problem for future me, we'll probably need further tweaks (kinda of a
hassle to dual-output with tsc, might need swc or some other transpiler
(not bundler).

---------

Co-authored-by: GitHub Actions Bot <>
Co-authored-by: developer-experience-bot[bot] <91079284+developer-experience-bot[bot]@users.noreply.github.com>
Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>