test(quantic): Tests for QuanticSearchBoxSuggestionsList (#5425)
## [SFINT-5645]

- Added unit tests for the QuanticSearchBoxSuggestionsList
- Added e2e tests for the QuanticSearchBoxSuggestionsList (leveraging
the exampleSearchbox page)
- Created a new ActionLocalStorage LWC to be used in tests to read/set a
value in the localStorage, this allows us to correctly handle the
localStorage even between the different versions of LockerService and
Lightning Web Security.
- Created a new QuerySuggestObject to use in tests to intercept and mock
calls to the /querysuggest API.
- Updated some other tests that were testing the same stuff I was
adding,
- for example, modified the QuanticSearchboxInput tests to remove the
tests that were testing the rendering of querySuggestions, to instead
test that the right data is passed to the
QuanticSearchBoxSuggestionsList component.

[SFINT-5645]:
https://coveord.atlassian.net/browse/SFINT-5645?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

### Unit tests


![image](https://github.com/user-attachments/assets/407a7029-d2f9-4043-8448-5874bc13c04c)
c

### E2E tests


![image](https://github.com/user-attachments/assets/e863ba0e-dd18-4559-87e2-7842bc0b5618)

> I am unsure about some of those e2e tests, if they are worth keeping.
> For example the input focus/blur,
> I am also questioning the worth of the tests that checks that we send
a search request with the right query when selecting a value in the
suggestions, that's pretty much covered across unit tests of the
component and also Headless unit tests.

For the tests that are skipped, this is because the Insight use-case
doesn't support the recent queries.

### Component to handle the local storage

- Has an option to specify a key for the localstorage and will display
its key as the label for the input.
- There's a button to set the value
- There's a button to read and print the value so you can test *after*
making modifications the resulting localstorage.


![image](https://github.com/user-attachments/assets/599ea629-af95-4ab9-9354-06c7799d7d7d)