fix(headless): missing actionCause analytics (#5436)
Adds the missing `actionCause` property to the analytics section sent
when calling the /generate endpoint.

[SVCC-5132](https://coveord.atlassian.net/browse/SVCC-5132)


# How to test

Change the `genqa.html` file to change from:

```
<atomic-generated-answer></atomic-generated-answer>
```

into:

```
<atomic-generated-answer answer-configuration-id="fc581be0-6e61-4039-ab26-a3f2f52f308f">
</atomic-generated-answer>
```

Start the project with `npx nx run atomic:dev` and load this url in your
browser:
http://localhost:3333/examples/genqa.html#q=how%20to%20resolve%20netflix%20connection%20with%20tivo.

Inspect results in DevTools network tab.

# Explanations

Pass the actionCause along with the existing analytics. 
The actionCause was not present in the state or the parameters where we
registered to the buildGeneratedAnswer.

As such this:
- Adds them to the Search state
- Creates a new search action to update the state
- Updates search-slice to handle change

[SVCC-5132]:
https://coveord.atlassian.net/browse/SVCC-5132?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ