feat(quantic): added attach results feature to the example insight panel (#5432)
## [SFINT-6119](https://coveord.atlassian.net/browse/SFINT-6119)

This PR introduces the "Attach Results to Case" feature to the example
Insight Panel in Quantic.

## Implementation Details
This feature enables attaching and detaching results to a case using
methods exposed by the AttachToCaseController Apex class from the [Coveo
for Salesforce (C4SF)
package](https://github.com/coveo-platform/salesforceintegration/blob/develop/force-app/main/default/classes/AttachToCaseController.cls).

To avoid introducing a direct dependency on the C4SF package in the
example project, we introduce a new service called AttachToCaseService:
[attachToCaseService.js](https://github.com/coveo/ui-kit/blob/52d8325f5f63476f0817a515961e07c75c9537d2/packages/quantic/force-app/solutionExamples/main/lwc/attachToCaseService/attachToCaseService.js)
This service provides stub implementations that are used in the example
Insight Panel.

In a production environment, integrators should replace this stub with
the real implementation provided in attachToCaseServiceProd.js:
`attachToCaseServiceProd.js`:
https://gist.github.com/mmitiche/32cdf65d66b6100b10f70cb47e434169

## Next Steps
We will ensure that this pattern and its usage are properly documented
to guide developers in customizing the Insight Panel for their
production needs.


## Demo



https://github.com/user-attachments/assets/87c29678-8795-475d-b282-fbb7bc905eae




[SFINT-6119]:
https://coveord.atlassian.net/browse/SFINT-6119?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: jpmarceau <39384459+jpmarceau@users.noreply.github.com>