fix(atomic): remove title from search box submit button (#5456)
https://coveord.atlassian.net/browse/KIT-4238

## Before
When using a voice over, when navigating to the submit button on the
search box, you heard : _"Search"_, _"Button"_, _"Search field with
suggestions. Suggestions may be available under this field. To send,
press Enter."_

The last part is wrong because suggestions are not available when
navigating to the submit button. This is text that is meant to be read
when focusing the `textarea`. This was found by @lvu285 during a qa
check on query suggestions :
https://github.com/coveo/ui-kit/pull/5205#issuecomment-2851475812

## After

By removing the title, the voice over now simply says : _"Search"_,
_"Button"_ which is much clearer.
You now hear the third part only when focusing the textarea.