chore: add support for using copilot prompts/instructions from atomic (#5552)
Right now, it is not possible to use copilot prompts/instructions if you
open a specific package, for example atomic, instead of the full ui-kit
repo.

With workspace settings inside a package, it is possible to set custom
paths to the prompts/instructions files

This PR adds this to the atomic package only. We can add the same file
to other packages later if needed.

https://coveord.atlassian.net/browse/KIT-4436