fix(headless commerce): ensure headless uses the same relay reference between engine and thunk (#5578)
https://coveord.atlassian.net/browse/KIT-4308

## Problem
The reference between engine.relay and the relay used in the
`createAsyncThunk` actions was desynced in commerce because of the line
in the commerce engine here:


https://github.com/coveo/ui-kit/blob/9e01022db1d190ed5f8c13fd720b83886a32a3f9/packages/headless/src/app/commerce-engine/commerce-engine.ts#L138

This made it so when you modified relay via engine.relay.updateConfig,
it would only change engine.relay and not extra.relay from the
`createAsyncThunk` actions.

## Fix
return a getter for relay pointing to the original engine.

I added tests for other engines even though the bug only happened in
commerce.