fix(atomic): fix "see all results" button deleting the query in the search box (#5587)
https://coveord.atlassian.net/browse/KIT-4355

This fix removes a useless line since every onClick suggestion must use
the search box controller to either trigger a search with the proper
query. There is no need for this line since the value of the text area
is the value in the controller state which is a lit state and is the one
changed by the suggestions. The value updates itself. This line caused a
race condition in the case of the show all results button since this one
had no query. Since it does not have a query. It would sometimes run
this line after the onSelect on the button thus removing the query. If
this line was done prior to the onSelect, it would work fine.

## Before 


https://github.com/user-attachments/assets/b3466201-d82e-41aa-a266-18b7ed382488



## After


https://github.com/user-attachments/assets/225efa10-a3bc-4845-a077-4e6649d4fad6