feat(cli): port over create-atomic from cli repo (#5623)
This is the first PR out of a suite of changes to alleviate issues with
Coveo CLI + Atomic related commands (mostly: Coveo CLI using out of date
version of Atomic and Headless).

First step is to bring over all "atomic template packages" from the CLI
to the UI-KIT repo. The general idea is that since they are now part of
the same mono repo, the version will be auto updated on each new release
of Atomic/Headless.

As well, we can better add build task to verify that the templates are
still working on each new version.

Its a big PR, but pretty much all of it is a dumb copy paste from the
CLI repo:
https://github.com/coveo/cli/tree/master/packages/ui/atomic/create-atomic

The goal of this PR is just to make sure that the build and release
process are working well.

I also applied the changes from this sample from @louis-bompart :
https://github.com/coveo/atomic-migration-v2-to-v3/pull/1/files

I will be porting over more atomic template related packages, and then
can test/update CLI commands to use this "new" process instead, so will
probably be updating those templates in subsequent PR.

## Tested by: 
* Manual publish of `@coveo/atomic-template` on npm with the
`1.40.9-test.1` version (--tag test)
* Run `npx @coveo/cli ui:create:atomic -v 1.40.9-test.1
my_dummy_project`
* Check that it build and run with Atomic v3.

## Version
Will probably want to bump to `@coveo/atomic-template` to v2 before
finishing all this... just to make sure.

[KIT-4508](https://coveord.atlassian.net/browse/KIT-4508)

[KIT-4508]:
https://coveord.atlassian.net/browse/KIT-4508?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ