chore(quantic): remove ckeditor (#5685)
We had issues a couple of times regarding jsdoc & this plugin.
After checking the files' and repo's content, I noticed that the latest
versions of CKEditor don't use jsdoc and have removed the project
altogether; it is still "usable" and maintained in a version branch.

Given how "simple" the file is... Let's simplify all this by removing
the dependency and taking the only JS file we do need.