fix(atomic): clear keyboard active descendant when search box input changes due to keypress (#5724)
https://coveord.atlassian.net/browse/KIT-4595

At the moment, if you select a suggestion through keyboard arrow
navigation, submit a query, and then modify the search box input (e.g.,
by pressing backspace, or by entering characters and/or whitespaces in
the input), pressing 'Enter' will not submit the query, because the
`keyboardActiveDescendant` will still be set to the ID of the suggestion
that was selected.

Clearing the keyboard active descendant fixes that.