fix(headless): fix ua events not firing with answer-api (#5719)
https://coveord.atlassian.net/browse/SVCC-5272

## The bug 

When using the answer-api no ua events were sent for streamingEnd,
like/dislike buttons or copy to clipboard action. This was because of
the way we were retrieving the data from the answerAPI request from the
state. Here is the step by step explanation of the origin of the bug:

1. In
[generated-answer-analytics-actions](https://github.com/coveo/ui-kit/blob/65c2238aff401a813cff3ec5c116938c1dc1c716/packages/headless/src/features/generated-answer/generated-answer-analytics-actions.ts#L353),
we need to have the rgaID to be able to make the analytics call, but
that `rgaID` is `undefined`
2. The reason it is `undefined` is because, when selecting the answer
from the Redux Toolkit Query, we pass the
`constructAnswerQueryParams(...)` to the selector. But there is a
volatile value in the answer query payload; the clientTimestamp.

Additionally, the fetch request also has access to the
[engine.navigatorContext](https://github.com/coveo/ui-kit/blob/65c2238aff401a813cff3ec5c116938c1dc1c716/packages/headless/src/controllers/knowledge/generated-answer/headless-answerapi-generated-answer.ts#L167)
and passes it along to the answer API payload builder. That
navigatorContext is then used to populate the analytics object. That
navigatorObject is not available from the
[selector](https://github.com/coveo/ui-kit/blob/65c2238aff401a813cff3ec5c116938c1dc1c716/packages/headless/src/features/generated-answer/generated-answer-selectors.ts#L19)
used in the [analytics actions
logger](https://github.com/coveo/ui-kit/blob/65c2238aff401a813cff3ec5c116938c1dc1c716/packages/headless/src/features/generated-answer/generated-answer-analytics-actions.ts#L352).

(The navigatorContext looks like: 
```
{
    "referrer": "http://localhost:3333/examples/genqa.html",
    "userAgent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/138.0.0.0 Safari/537.36",
    "location": "http://localhost:3333/examples/genqa.html#q=how%20to%20resolve%20netflix%20connection%20with%20tivo",
    "clientId": "2667c24d-449c-40ee-a9cd-2ea702b6fab8"
}
```
)
3. Since the body we pass to the select RTK query is not the same as the
body that was serialized to the redux state, the select is unable to
access any data and thus, the `rgaID` is null.

## The solution

When we are using RTK query to select (and not to fetch), we remove the
analytics object from the selector body, and we also do the same thing
for the [serializeQueryArgs
method](https://github.com/coveo/ui-kit/blob/65c2238aff401a813cff3ec5c116938c1dc1c716/packages/headless/src/api/knowledge/stream-answer-api.ts#L201).
This means that the key in the store does not contain any analytics, but
since we don't need analytics in the select, it's not a problem.

## How to test

In genqa.html
1. Add an answer config ID:

```
          <atomic-generated-answer answer-configuration-id="fc581be0-6e61-4039-ab26-a3f2f52f308f">
          </atomic-generated-answer>
```
2.  Initialize in legacy mode:

```
      // Initialization
      await searchInterface.initialize({
        accessToken: 'xx564559b1-0045-48e1-953c-3addd1ee4457',
        organizationId: 'searchuisamples',
        search: {
          pipeline: 'genqatest',
        },
        analytics: {analyticsMode: 'legacy'},
      });
```

Do some actions in the genQA tab, you should see the custom events be
sent:


https://github.com/user-attachments/assets/78cad25f-6bc5-4af8-84b7-9001f1f7d8a9